Update Prometheus to scrape data and control plane
The existing telemetry pipeline relies on Prometheus scraping the
Telemetry service, which will soon be removed.

This change configures Prometheus to scrape the conduit proxies directly
for telemetry data, and the control plane components for control-plane
health information. This affects the output of both `conduit install`
and `conduit inject`.

Fixes #428, #501

Signed-off-by: Andrew Seigner <siggy@buoyant.io>